Mejora en el script de instalación y actualización de Virtual Display
- Se agrega verificación de existencia de directorio "generiz" antes de mover archivos y crear la tarea programada.
- Se añade la eliminación y reinstalación del Virtual Display al script de tarea que se ejecuta en cada reinicio.
- Se incluyen colores y una barra de progreso en el entorno del script principal.
- Se asegura que la tarea programada se ejecute con privilegios de administrador.

+++ solucion.bat
@@ -1,27 +1,37 @@
 @echo off
 TITLE Powered by generiz
 color 80
-MKDIR c:\generiz
-MOVE carpeta\Win32 c:\generiz\
-MOVE carpeta\x64 c:\generiz\
-MOVE carpeta\usbmmidd.bat c:\generiz\
-MOVE carpeta\usbmmidd.cat c:\generiz\
-MOVE carpeta\usbmmidd.inf c:\generiz\
-MOVE carpeta\deviceinstaller.exe c:\generiz\
-MOVE carpeta\deviceinstaller64.exe c:\generiz\
-MOVE carpeta\tarea.bat c:\generiz\
 
-@echo Instalado drivers...
+IF EXIST "c:\generiz\" (
+    echo "generiz ya está instalado."
+) ELSE (
+    MKDIR c:\generiz
+    MOVE carpeta\Win32 c:\generiz\
+    MOVE carpeta\x64 c:\generiz\
+    MOVE carpeta\usbmmidd.bat c:\generiz\
+    MOVE carpeta\usbmmidd.cat c:\generiz\
+    MOVE carpeta\usbmmidd.inf c:\generiz\
+    MOVE carpeta\deviceinstaller.exe c:\generiz\
+    MOVE carpeta\deviceinstaller64.exe c:\generiz\
+    MOVE carpeta\tarea.bat c:\generiz\
 
+    @echo Instalando drivers...
+    cd c:\generiz\
+    deviceinstaller64 install usbmmidd.inf usbmmidd
+    deviceinstaller64 enableidd 1
 
-cd c:\generiz\
-deviceinstaller64 install usbmmidd.inf usbmmidd
-deviceinstaller64 enableidd 1
+    @echo Creando tarea...
+    schtasks /create /tn generiz /tr c:\generiz\tarea.bat /sc onstart /RL HIGHEST
 
-pause
-@echo creando tarea...
+    pause
+)
 
-
-schtasks /create /tn generiz /tr c:\generiz\tarea.bat /sc onstart
+:progress
+color 07
+echo.
+echo Por favor espere mientras se configura el entorno...
+echo  ██████████████████████████ 100%%
+echo.
+color 80
 
 pause
